third_party/blink/web_tests/external/wpt/html/browsers/the-window-object/garbage-collection-and-browsing-contexts/discard_iframe_history_3-expected.txt

This is a testharness.js-based test.
FAIL Removing iframe from document removes it from history assert_array_equals: expected property 2 to be 1 but got 2 (expected array [1, 2, 1] got [1, 2, 2])
Harness: the test ran to completion.
